refactor: Coordinator・CenterCommandをClusterMapViewから分離 (#167)
